requirements.txt
flask-script==2.0.6
Flask-Testing==0.6.2
gunicorn==19.7.1
flask-cors==3.0.2
dateutils==0.6.6
flask-restplus==0.11.0
pybind11==2.2.3
cysignals==1.7.1
cython==0.28.3